feat: onTNAProductClick Deprecate, handler를 슬롯 컴포넌트 내부로 이동
