+++ packages/triple-document/src/triple-document.tsx
@@ -25,7 +25,6 @@ export function TripleDocument({
   onResourceClick,
   onImageClick,
   onLinkClick,
-  onTNAProductClick,
   imageSourceComponent,
   deepLink,
   cta,
@@ -90,7 +89,6 @@ export function TripleDocument({
                               onResourceClick: resourceClickHandler,
                               onImageClick,
                               onLinkClick: linkClickHandler,
-                              onTNAProductClick,
                               ImageSource: imageSourceComponent,
                               deepLink,
                               videoAutoPlay,
